Update to web-ext-plugin version 2.x.x
Required addional compiler options for ts-node to be able to handle ESM
modules, which web-ext and web-ext-plugin now use. Also Webpack
configuration needed to be split for ESLint to still be able to use it.
See https://github.com/webpack/webpack-cli/issues/2458 for details
regarding ts-node and ESM.
