+++ package.json
@@ -121,7 +121,7 @@
     "ts-node": "^10.8.1",
     "ts-prune": "^0.10.3",
     "typescript": "^4.7.3",
-    "web-ext-plugin": "^1.8.2",
+    "web-ext-plugin": "^2.1.0",
     "webpack": "^5.73.0",
     "webpack-cli": "^4.10.0"
   },
